requirements.txt
absl-py==0.6.1
astor==0.7.1
audioread==2.1.6
boto3==1.9.83
botocore==1.12.83
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
console-progressbar==1.1.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
dill==0.2.8.2
dlib==19.16.0
docutils==0.14
face-recognition==1.2.3
face-recognition-models==0.3.0
fire==0.1.3
gast==0.2.0
grpcio==1.17.1
h5py==2.8.0
idna==2.8
imageio==2.4.1
imutils==0.5.2
jmespath==0.9.3
joblib==0.13.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.26.0
Markdown==3.0.1
matplotlib==3.0.2
moviepy==0.2.3.5
msgpack==0.6.0
msgpack-numpy==0.4.4.2
mysql-connector==2.1.6
networkx==2.2
nltk==3.4
numba==0.41.0
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
Pillow==5.3.0
pocketsphinx==0.1.15
protobuf==3.6.1
psutil==5.4.8
public==2018.11.20
pyparsing==2.3.0
python-dateutil==2.7.5
python-rake==1.4.5
pytube==9.3.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
requests==2.21.0
resampy==0.2.1
s3transfer==0.1.13
schedule==0.6.0
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
slidingwindow==0.0.13
SpeechRecognition==3.8.1
tabulate==0.8.2
tensorboard==1.12.1
tensorflow-gpu==1.12.0
tensorpack==0.9.0.1
termcolor==1.1.0
textblob==0.15.2
tf-pose==0.1.1
toolz==0.9.0
torch==0.4.0
torchvision==0.2.1
tqdm==4.28.1
urllib3==1.24.1
values==2018.11.22
webrtcvad==2.0.10
Werkzeug==0.14.1
